% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{aggregate_ww}
\alias{aggregate_ww}
\title{Aggregate the WW data to the state-level}
\usage{
aggregate_ww(
  ww_data_raw,
  forecast_date,
  calibration_time,
  hosp_reporting_delay
)
}
\arguments{
\item{ww_data_raw}{The site leve ww data after being cleaned}

\item{forecast_date}{date forecast is being made}

\item{calibration_time}{duration of calibration period for hospitalizations}

\item{hosp_reporting_delay}{the hospital reporting delay (how much is
being nowcasted)}
}
\value{
a dataframe with wastewater data for all locations,
the weekly avg pcr concentration, and the date
}
\description{
Aggregate the WW data to the state-level
}
