% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EDA_utils.R
\name{boxplot_whiskers}
\alias{boxplot_whiskers}
\title{Get the upper and lower ends of the whiskers of a single boxplot of data in x}
\usage{
boxplot_whiskers(x)
}
\description{
Get the upper and lower ends of the whiskers of a single boxplot of data in x
}
