% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EDA_utils.R
\name{categorical_entropy}
\alias{categorical_entropy}
\title{Entropy of sample from categorical distribution}
\usage{
categorical_entropy(x, cats = NA)
}
\description{
Entropy of sample from categorical distribution
}
