% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfaforecastrenewalww-package.R
\docType{package}
\name{cfaforecastrenewalww-package}
\alias{cfaforecastrenewalww}
\alias{cfaforecastrenewalww-package}
\title{cfaforecastrenewalww: Wastewater informed COVID-19 hospital admissions forecasting}
\description{
Put something here
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/cdcgov/wastewater-informed-covid-forecasting/}
  \item Report bugs at \url{https://github.com/cdcgov/wastewater-informed-covid-forecasting/issues/}
}

}
\author{
\strong{Maintainer}: Kaitlyn Johnson \email{uox1@cdc.gov}

Authors:
\itemize{
  \item Sam Abbott \email{contact@samabbott.co.uk} (\href{https://orcid.org/0000-0001-8057-8037}{ORCID})
  \item Zachary Susswein \email{utb2@cdc.gov}
  \item Andrew Magee \email{rzg0@cdc.gov}
  \item Dylan Morris \email{dylan@dylanhmorris.com}
  \item Scott Olesen \email{ulp7@cdc.gov}
}

}
\keyword{internal}
