% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_model.R
\name{compile_model}
\alias{compile_model}
\title{This code was adapted from code written
(under an MIT license) as part of the `epinowcast`
package (https://github.com/epinowcast/epinowcast)
Compile a stan model while pointing at the package default
include directory (`stan`) for #include statements}
\usage{
compile_model(
  model_filepath,
  include_paths = system.file("stan", package = "cfaforecastrenewalww"),
  threads = FALSE,
  target_dir = tempdir(),
  stanc_options = list(),
  cpp_options = list(),
  verbose = TRUE,
  ...
)
}
\arguments{
\item{model_filepath}{path to .stan file defining the model}

\item{include_paths}{path(s) to directories to search for files
specified in #include statements. Passed to [cmdstanr::cmdstan_model()].
Defaults to the `stan` subdirectory of the installed
`cfaforecastrenewalww` package.}

\item{threads}{Number of threads to use in model compilation,
as an integer. Passed to [cmdstanr::cmdstan_model()].
Default `FALSE` (use single-threaded compilation).}

\item{target_dir}{Directory in which to save the compiled
stan model binary. Passed as the `dir` keyword argument to
[cmdstanr::cmdstan_model()]. Defaults to a temporary directory
for the R sessions (the output of [tempdir()]).}

\item{stanc_options}{Options for the stan compiler passed to
[cmdstanr::cmdstan_model()], as a list. See that function's
documentation for more details. Default `list()` (use default
options).}

\item{cpp_options}{Options for the C++ compiler passed to
[cmdstanr::cmdstan_model()], as a list. See that function's
documentation for more details. Default `list()` (use default
options).}

\item{verbose}{Write detailed output to the terminal while
executing the function? Boolean, default `TRUE`.}

\item{...}{Additional keyword arguments passed to
[cmdstanr::cmdstan_model()].}
}
\value{
The resulting `cmdstanr` model object, as the output
of [cmdstanr::cmdstan_model()].
}
\description{
This function reads in and optionally compiles a Stan model.
It is written to search the installed package `stan` directory
for additional source files to include. Within each stan file,
use #include {path to your file with the `stan` directory}.stan
}
\details{
compile_model
}
