% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_to_logmean}
\alias{convert_to_logmean}
\title{Convert to logmean in lognorm distribution}
\usage{
convert_to_logmean(mean, sd)
}
\arguments{
\item{sd}{}
}
\value{
corresponding mean of the lognormal distribution
}
\description{
see arithmetic moments here
https://en.wikipedia.org/wiki/Log-normal_distribution
}
