% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_to_logsd}
\alias{convert_to_logsd}
\title{Convert to logsd in lognormal distribution}
\usage{
convert_to_logsd(mean, sd)
}
\arguments{
\item{sd}{}
}
\value{
corresponding stdev of the lognormal distribution
}
\description{
@description see arithmetic moments here
https://en.wikipedia.org/wiki/Log-normal_distribution
}
