% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_dir}
\alias{create_dir}
\title{Create a new directory if one doesn't exist}
\usage{
create_dir(output_file_path)
}
\arguments{
\item{output_file_path}{file path that may or may not need to be created}
}
\description{
Function to create a directory for the specified output file path if needed.
dir_create won't throw a warning if its already made though!
}
