% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_predictive_analysis.R
\name{.do_1_pps_analysis}
\alias{.do_1_pps_analysis}
\title{Internal helper function to do a single analysis of PPS data}
\usage{
.do_1_pps_analysis(i, model, real_data_draws, stan_predictive_data, ci, ...)
}
\arguments{
\item{i}{index of simulated data to analyze
the parameters from which the data was simulated}

\item{model}{CmdStanModel to fit}

\item{real_data_draws}{draws object containing MCMC samples from real-data model fit}

\item{stan_predictive_data}{list of simulated datasets, in same order as real_data_draws}

\item{ci}{length-2 numeric vector of [lower, upper] values for CI}

\item{...}{further arguments passed to model$sample(), e.g. # chains}
}
\value{
a summary of the performance of the
}
\description{
Takes in index, parameter values from which data was simulated,
list of simulated data, the model to run, and other arguments for either
the model running or summarizing.
Runs stan on the simulated data with this index returns a summary of performance.
}
\keyword{internal}
