% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_model_outputs.R
\name{.get_1_param}
\alias{.get_1_param}
\title{Internal helper for get_param_samples_long_df}
\usage{
.get_1_param(param_info, all_draws)
}
\arguments{
\item{param_info}{single-row data frame (or tibble), originating from get_model_param_df().}

\item{all_draws}{MCMC samples as a draws_array object}
}
\value{
data frame with parameter name, values, and (as applicable) information
about the spatiotemporal aspects of the parameter
}
\description{
This function makes a "long" data frame for a single parameter. This may be a
scalar-, vector-, or matrix-valued parameter.
}
\keyword{internal}
