% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_model_outputs.R
\name{.get_param_spreader_strings}
\alias{.get_param_spreader_strings}
\title{Internal helper for .get_1_param}
\usage{
.get_param_spreader_strings(param_info)
}
\arguments{
\item{param_info}{single-row data frame (or tibble), originating from get_model_param_df().}
}
\value{
named vector,
}
\description{
Provides information required to spread() vector- or matrix-valued parameters.
}
\keyword{internal}
