% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_pps_data.R
\name{.make_1_pps_data}
\alias{.make_1_pps_data}
\title{Internal helper for make_pps_data()}
\usage{
.make_1_pps_data(pps_hosp_data, pps_ww_data, real_data, config)
}
\arguments{
\item{pps_hosp_data}{one draw's worth of pps hospital data}

\item{pps_ww_data}{one draw's worth of pps ww data}

\item{real_data}{the stan data object}

\item{config}{the output of get_config_vals() for the original (real) data}
}
\value{
a list, as real_data but with the pps hospital and ww data instead
of the real observations
}
\description{
Takes in pre-processed predictive hospital and wastewater data, plus a real-data
stan input and a config. Outputs a posterior predictive dataset.
}
\keyword{internal}
