% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_predictive_analysis.R
\name{.summarize_performance}
\alias{.summarize_performance}
\title{Internal helper function to summarize how well an analysis of posterior
predictive data performed}
\usage{
.summarize_performance(true_param_draws, fit_draws, ci)
}
\arguments{
\item{true_param_draws}{draws-family object containing a single draw with
the parameters from which the data was simulated}

\item{fit_draws}{draws-family object containing the MCMC}

\item{ci}{length-2 numeric vector of \link{lower, upper} values for CI}
}
\value{
a list, each of which is a list of formatted stan data
}
\description{
Takes in true parameter values and MCMC samples and computes summaries of the
samples, possibly wrt the truth.
}
\keyword{internal}
