% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_df}
\alias{example_df}
\title{Example input data set}
\format{
## example_df
A data frame with 635 rows and 13 columns
\describe{
  \item{t}{Time}
  \item{lab_wwtp_unique_id}{Unique identifier for each unique combination
        of sampling site and testing lab}
  \item{log_conc}{Genome copied per liter, log10}
  ETC.
}
}
\usage{
example_df
}
\description{
SOME HIGH-LEVEL DESCRIPTION OF THE DATA. MAYBE JUST POINT READERS TO
THE FUNCTION THAT GENERATES IT?
}
\keyword{datasets}
