% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_params.R
\docType{data}
\name{example_params}
\alias{example_params}
\title{Example model parameters and hyperparameters for priors}
\format{
A .toml file that gets formatted as a 1 row dataframe where each column is a
parameter name and its value is indicated in the first row
\describe{
\item{\code{uot}}{"unoubserved time" an integer that dictates the duration
of time (in days) that the model is initialized with exponential growth
before any observations are available.}
\item{\code{r_prior_mean}}{Prior on mean R(t)}
\item{\code{r_prior_sd}}{Prior on standard deviation in the mean R(t)}
\item{\code{sigma_rt_prior}}{Prior on the standard deviation between
subpopulation level R(t)s}
\item{\code{i0_certainty}}{Prior on the certainty in initial infections
I0/N, representing the effective number of binomial trials in the beta
prior centered on I0/N}
\item{\code{initial_growth_prior_mean}}{Prior on mean exponential growth
rate(daily) in the unobserved time during infection initialization}
\item{\code{initial_growth_prior_sd}}{Prior on the standard deviation in
the daily exponential growth rate during the unobserved time}
\item{\code{autoreg_rt_a}}{Prior on the autoregulation term on the R(t)
trend and the AR(1) process in subpopulation level R(t), shape1 parameter of
a beta prior}
\item{\code{autoreg_rt_b}}{Prior on the autoregulation term on the R(t)
trend and the AR(1) process in subpopulation level R(t), shape2 parameter of
a beta prior}
\item{\code{eta_sd_sd}}{Prior on standard deviation in the deviation in the
R(t) trend}
\item{\code{infection_feedback_prior_logmean}}{Prior on the log mean
infection feedback term}
\item{\code{infection_feedback_prior_logsd}}{Prior on log standard
deviation of the infection feedback term)}
\item{\code{p_hosp_mean}}{Prior on the mean infection hospital admissions
rate}
\item{\code{p_hosp_sd_logit}}{Prior on the logit scale of the standard deviation
of the mean infection hospital admissions rate}
\item{\code{p_hosp_sd_sd}}{Prior on the standard deviation in the time-varying
deviation in the infection hospital admissions rate}
\item{\code{inv_sqrt_phi_prior_mean}}{Prior on the mean of the inverse square
root of the negative binomial \code{phi} of the hospital admissions
observation process}
\item{\code{inv_sqrt_phi_prior_sd}}{Prior on the standard deviation of the
inverse square root of the negative binomial \code{phi} of the hospital
admissions observation process}
\item{\code{wday_effect_prior_mean}}{Prior on the mean amount of weight
assigned to each day of the week proportional to the whole week}
\item{\code{wday_effect_prior_sd}}{Prior on the standard deviation in the
amount of weight assigned to each day of the week proportional to the whole
week}
\item{\code{ml_of_ww_per_person_day}}{Set value of the estimated number
of mL of wastewter produced per person per day}
\item{\code{t_peak_mean}}{Prior on the mean time of peak fecal shedding in
wastewater}
\item{\code{t_peak_sd}}{Prior on the standard deviation in the time of peak
fecal shedding in wastewater}
\item{\code{viral_peak_mean}}{Prior on mean peak viral load (in log10 scale)
of fecal shedding of viral genomes in wastewaster}
\item{\code{viral_peak_sd}}{Prior on the standard deviation in the peak viral
load (in log10 scale) of fecal shedding of viral genomes in wasteawter}
\item{\code{duration_shedding_mean}}{Prior on mean duration of fecal shedding
in wastewater}
\item{\code{duration_shedding_sd}}{Prior on the standard deviation of
fecal shedding in wastewater}
\item{\code{log10_g_prior_mean}}{Prior on the mean number of genomes shed
per infected individual in log10 scale}
\item{\code{log10_g_prior_sd}}{Prior on standard deviation in the number of
genomes shed per infected individual in log10 scale}
\item{\code{log_g_prior_mean}}{Prior on the mean number of genomes shed
per infected individual in log scale}
\item{\code{log_g_prior_sd}}{Prior on standard deviation in the number of
genomes shed per infected individual in log scale}
\item{\code{sigma_ww_site_prior_mean_mean}}{Prior on mean of the mean
site-lab-level observation error}
\item{\code{sigma_ww_site_prior_mean_sd}}{Prior on standard deviation of the
mean site-lab-level observation error}
\item{\code{sigma_ww_site_prior_sd_mean}}{Prior on the mean of the standard
deviation across sites in site-lab-level observation errors}
\item{\code{sigma_ww_site_prior_sd_sd}}{Prior on standard deviaiton of the
standard deviation across sites in site-lab level observation errors}
\item{\code{ww_site_mod_sd_sd}}{Prior on the standard deviation in the
standard deviation between site-lab level multipliers}
\item{\code{log_phi_g_prior_mean}}{Prior on mean individual level dispersion
in number of genomes shed per infection, not currently used in the model}
\item{\code{log_phi_g_prior_sd}}{Prior on the standard deviation in
individual level dispersion in number of genomes shed per infection, not
currently used in the model}
\item{\code{mu_gi}}{Set parameter for the log mean of the generation
interval}
\item{\code{sigma_gi}}{Set parameter for the log standard deviaiton of the
generation interval}
\item{\code{gt_max}}{Integer indicating the maximum number of days
for which secondary transmission can occur}
\item{\code{r}}{Exponential growth rate on the correction factor applied to
the incubation period estimate}
\item{\code{backward_shape}}{Set parameter for the backward shape parameter
of the Weibull distribution for the incubation period}
\item{\code{backward_scale}}{Set parameter for the backward scale parameter
of the Weibull distribution for the incubation period}
\item{\code{neg_binom_mu}}{Set parameter for the mean of the negative
binomially distributed time from symptom onset to hospital admission}
\item{\code{neg_binom_size}}{Set parameter for the size parameter of the
negative binomially distributed time from symptom onset to hospital
admissions}
}
}
\source{
The set parameters for the generation interval and incubation period were
obtained from:
Park, Sang Woo, et al. \href{https://doi.org/10.1073/pnas.2221887120}{"Inferring the differences in incubation-period and generation-interval distributions of the Delta and Omicron variants of SARS-CoV-2."}
Proceedings of the National Academy of Sciences 120.22 (2023):
e2221887120.

See \href{https://github.com/CDCgov/wastewater-informed-covid-forecasting/blob/prod/model_definition.md#model-parameters}{\verb{Model Parameters}}
for the full set of references for each prior.
}
\description{
The \code{example_params.toml} file contains modifiable model inputs that get passed
to stan to fit the model. The parameters used here have been chosen by the
authors to best reflect SARS-CoV-2 infection, hospital admissions, and
wastewater components. See the
\href{https://github.com/CDCgov/wastewater-informed-covid-forecasting/blob/prod/model_definition.md#model-parameters}{\verb{Model Parameters}}
section of the model_definition.md in our Github for more information/justification of these
prior/parameter choices.

If running the model on SARS-CoV-2 hospital admissions and wastewater concentrations,
it may be reasonable to use these example parameters as a starting place for your
data (aka none are dataset specific). The values that will need tweaking for a
specific dataset are defined in the documentation for \link{example_df}.

\if{html}{\out{<div class="sourceCode">}}\preformatted{example_params <- cfaforecastrenewalww::get_params(
  fs::path_package(
    "extdata", "example_params.toml",
    package = "cfaforecastrenewalww"
  )
)
}\if{html}{\out{</div>}}
}
\details{
These parameters specify the priors on key parameters, except where indicated
as set parameters. In these instances (generation interval, incubation period,
symptom onset to hospital admissions) we pre-define the delay distrubions using
estimates from the literature.
}
\keyword{datasets}
