% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{fit_site_level_model}
\alias{fit_site_level_model}
\title{Fit the stan model to a single slice of data}
\usage{
fit_site_level_model(
  train_data,
  params,
  model_file,
  forecast_date,
  run_id,
  date_run,
  forecast_time,
  generation_interval,
  inf_to_hosp,
  infection_feedback_pmf,
  model_type,
  output_file_path,
  include_hosp = 1,
  compute_likelihood = 1,
  n_draws = 100,
  n_chains = 4,
  iter_sampling = 500,
  iter_warmup = 250,
  n_parallel_chains = 4,
  write_files = TRUE,
  output_full_df = FALSE,
  ...
)
}
\arguments{
\item{train_data}{for a single location}

\item{params}{of the model, these should be the same across all runs}

\item{model_file}{compiled stan model (an upstream target)}

\item{config_vars}{tells the model whether or not to include WW,
include hosp, compute likelihood, and what damp type.}
}
\value{
a dataframe with 100 model draws for all time points for WW
 concentraiton, hospitalizaitons, hosp per 100k, and the matched data
}
\description{
Fit the stan model to a single slice of data
}
