% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_model_outputs.R
\name{format_hosp_draws_hub_sub}
\alias{format_hosp_draws_hub_sub}
\title{Prepare posterior draws of hospital admissions for the COVID-19 forecast hub}
\usage{
format_hosp_draws_hub_sub(hosp_draws)
}
\arguments{
\item{hosp_draws}{dataframe containing the posterior draws from
all time points of the hospital admissions for a single model and location}
}
\value{
a dataframe formatted for COVID-19 forecast hub submission
}
\description{
Prepare posterior draws of hospital admissions for the COVID-19 forecast hub
}
