% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_model_outputs.R
\name{get_all_quantiles}
\alias{get_all_quantiles}
\title{Get all quantiles}
\usage{
get_all_quantiles(model_draws)
}
\arguments{
\item{model_draws}{}
}
\value{
a long dataframe of all the quantiled values of the model variables
columns names include all the things we had in the draws_df + metadata:
name, t, value, draw, quantile,
include_ww, forecast_date, hosp_reporting_delay,
location, data_type, date, ww, daily_hosp_admits,
daily_hosp_admits_for_eval, pop, obs_data, period, damp_type
}
\description{
Takes in the model draws and summarizes over them to get quantiles that
correspond to the quantiles reported in the COVID-19 Forecast Hub that
are used for generating WIS scores.
}
