% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{get_all_training_data}
\alias{get_all_training_data}
\title{Get long tibble of data needed for model fitting}
\usage{
get_all_training_data(
  ww_data_raw,
  forecast_date,
  location,
  calibration_time,
  geo_type,
  pull_from_local,
  hosp_data_dir,
  population_data_path,
  forecast_time,
  hosp_reporting_delay,
  ww_geo_type,
  hosp_data_source,
  include_ww,
  train_data_dir,
  model_type,
  write_files = TRUE,
  ...
)
}
\arguments{
\item{include_ww}{}
}
\value{
a long stacked dataframe with columns corresponding to model specs
and data specs
}
\description{
Get long tibble of data needed for model fitting
}
