% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{get_combo_quantile_plot}
\alias{get_combo_quantile_plot}
\title{Title}
\usage{
get_combo_quantile_plot(
  df,
  figure_file_path,
  days_to_show_forecast = 14,
  from_full_df = FALSE,
  write_files = TRUE
)
}
\arguments{
\item{df}{dataframe of either quantiles or the path to load in quantiles}

\item{figure_file_path}{directory to save figure}

\item{days_to_show_forecast}{duration of forecast to plot, default is 14}

\item{from_full_df}{if TRUE then df is quantiles object, if FALSE then df is a
dataframe of filepaths to load in the parameter object}

\item{write_files}{if TRUE then write files to specified location, if FALSE
then don't}
}
\value{
plot of bar chart of hospitalizations with mean state level
wastewater overlaid
}
\description{
Title
}
