% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_config_vals.R
\name{get_config_vals}
\alias{get_config_vals}
\title{Read in the config yaml}
\usage{
get_config_vals(path_to_config)
}
\arguments{
\item{path_to_config}{path the config yaml}
}
\description{
Read in the config yaml
}
