% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_model_outputs.R
\name{get_df_for_hub_submission}
\alias{get_df_for_hub_submission}
\title{Load in future hosp draws and combine into a single dataframe}
\usage{
get_df_for_hub_submission(
  df_of_filepaths,
  submitting_model_name,
  submission_file_path,
  repo_file_path,
  prod_run,
  write_files = TRUE,
  ...
)
}
\arguments{
\item{df_of_filepaths}{filepaths where the future hosp draws live}

\item{submitting_model_name}{name of the model that we will set as the
folder name}

\item{write_files}{whether or not to write to the forecast submission
file path}

\item{forecast_submission_filepath}{where we want the csv to be saved}
}
\value{
df of hub submission for all locations
}
\description{
Load in future hosp draws and combine into a single dataframe
}
