% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_model_outputs.R
\name{get_diagnostics}
\alias{get_diagnostics}
\title{Get a joint dataframe of data and model fit diagnostics}
\usage{
get_diagnostics(
  stan_fit_object,
  train_data,
  location,
  model_type,
  forecast_date,
  ...
)
}
\arguments{
\item{stan_fit_object}{The output of fit$sample() in cmdstanR}

\item{train_data}{The datarframe containing hospital admissions and
wastewater data for callibration of the model}

\item{location}{The location the model is being run on}

\item{model_type}{The type of model e.g. site-level observation error}

\item{forecast_date}{The date of the forecast}

\item{...}{}
}
\value{
Returns a dataframe containing diagnostic summaries
}
\description{
This returns a 10 row dataframe corresponding to 6 flags of the hospital
admissions or wastewater data and 4 rows corresponding to stan fit
diagnostics.
}
