% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_config_vals.R, R/process_model_outputs.R
\name{get_figure_file_path}
\alias{get_figure_file_path}
\title{Get the figure file path}
\usage{
get_figure_file_path(output_file_path, forecast_date, date_run)

get_figure_file_path(output_file_path, forecast_date, date_run)
}
\arguments{
\item{date_run}{}
}
\description{
Get the figure file path

Hardcoded figure filepath for figures per the read me
}
