% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_output_subdirectory_paths.R
\name{get_figure_output_subdirectory}
\alias{get_figure_output_subdirectory}
\title{Get the path to the subdirectory in which to
save figures for a given top-level output directory}
\usage{
get_figure_output_subdirectory(output_dir)
}
\arguments{
\item{output_dir}{Path to the output directory}
}
\value{
path to the figure subdirectory
}
\description{
Get the path to the subdirectory in which to
save figures for a given top-level output directory
}
