% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_model_outputs.R
\name{get_full_param_distrib}
\alias{get_full_param_distrib}
\title{Get full parameter distribution}
\usage{
get_full_param_distrib(
  all_draws,
  static_params,
  vector_params,
  matrix_params = NA
)
}
\arguments{
\item{all_draws}{draws from the stan object}

\item{static_params}{character vector of parameters that are static and
don't need any indexing}

\item{weekly_params}{character vector of parameters that are indexed by week}

\item{daily_ww_params}{character vector of parameters that are indexed by
day (t) and ww lab site}

\item{ww_lab_site_params}{character vector of parameters that are indexed by
unique combo of the wastewater lab and site}

\item{day_of_week_params}{character vector of parameters that are indexed by
day of week}
}
\value{
a dataframe containing the posterior distribution of each parameter
With indices corresponding to the parameters that have them
}
\description{
This function takes in vectors of parameters, grouped by their indexing,
and returns a tidy dataframe with all the draws of all the parameters and
the corresponding indices
}
