% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_model_outputs.R
\name{get_gen_quants_draws}
\alias{get_gen_quants_draws}
\title{Get generated quantities draws site level model}
\usage{
get_gen_quants_draws(all_draws, model_type)
}
\arguments{
\item{train_data}{}
}
\value{
A long dataframe with model draws from pred_hosp, pred_ww, R(t) and
p_hosp(t), where all but pred_ww have one value per time point and draw. Bc
we need all the column names to be the same, site_index is left empty for those.
The resulting dataframe has the following columns:
name, site_index, t, value, draw
}
\description{
This function takes the raw stan output from the site leve model and
arranges them in a tidy data format, with time and site_index as indexing
variables.
}
