% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delay_distribs.R
\name{get_generation_interval_gamma}
\alias{get_generation_interval_gamma}
\title{Get the generation interval from a gamma distribution}
\usage{
get_generation_interval_gamma(mean_gi, stdev_gi, gt_max)
}
\arguments{
\item{gt_max}{}
}
\value{
a normalized generation interval assuming gamma distribution
}
\description{
Get the generation interval from a gamma distribution
}
