% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delay_distribs.R
\name{get_generation_interval_lnorm}
\alias{get_generation_interval_lnorm}
\title{Get the generation interval from a lognormal distribution}
\usage{
get_generation_interval_lnorm(mu = 0.92877, sigma = 0.526, gt_max = 15)
}
\arguments{
\item{gt_max}{}
}
\value{
a normalized generation interval assuming a lognormal distribution
}
\description{
Get the generation interval from a lognormal distribution
}
