% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{get_hosp_data}
\alias{get_hosp_data}
\title{Gets hospitalization data that model will be trained on}
\usage{
get_hosp_data(
  hosp_data_source,
  geo_type,
  forecast_date,
  hosp_data_dir,
  population_data_path,
  pull_from_local = FALSE
)
}
\arguments{
\item{hosp_data_source}{Source of hospitalization data
from which to pull. Options include "NHSN" and "covidcast".}

\item{geo_type}{type of geographic region for which to pull data}

\item{forecast_date}{date for which we are producing a forecast}

\item{hosp_data_dir}{if pulling from local, where to get the data from}

\item{population_data_path}{path to a table of state populations}

\item{pull_from_local}{if TRUE, look for local files first,
default value is FALSE}
}
\value{
a data frame at the geo_type level of number of daily hospital admits
}
\description{
Gets hospitalization data that model will be trained on
}
