% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_init_fun}
\alias{get_init_fun}
\title{Get initialization function for stan}
\usage{
get_init_fun(model_type, train_data, params, stan_data)
}
\value{
the initialization function
}
\description{
Get initialization function for stan
}
