% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_model_outputs.R
\name{get_loc_model_map}
\alias{get_loc_model_map}
\title{Get location to model map}
\usage{
get_loc_model_map(
  df_of_filepaths,
  hosp_only_states,
  us_model_type = "state-level aggregated wastewater"
)
}
\arguments{
\item{df_of_filepaths}{the dataframe that comes out of fit model with
columns indicating the paths to the specified models and the model type
of those outputs}

\item{hosp_only_states}{the states we want to manually replace with the
hospital admissions only model for submission}

\item{us_model_type}{The default model type for the US}
}
\description{
Get location to model map
}
