% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_metadata}
\alias{get_metadata}
\title{Get metadata from a single group training data}
\usage{
get_metadata(train_data)
}
\arguments{
\item{train_data}{single group training data
(single location, model type, forecast date)}
}
\value{
single row dataframe containing values for forecast date, pop,
include_ww, hosp_reporting delay, and location
}
\description{
From a single model run, get the relevant metadata
}
