% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_output_metadata.R
\name{get_metadata_yaml}
\alias{get_metadata_yaml}
\title{Create YAML formatted metadata file
listing states without wastewater data,
states with insufficient wastewater data,
and states for which we used the hospital
admissions only model.}
\usage{
get_metadata_yaml(
  data_diagnostics_df,
  hosp_only_states,
  exclude_states,
  output_dir = NULL,
  prod_run = FALSE
)
}
\arguments{
\item{data_diagnostics_df}{Data frame of data
diagnostics by location}

\item{hosp_only_states}{States for which we chose
to use the hospitalization only model, as a vector
with states coded identically to how they are in
the `location` column of `full_diagnostics_df`}

\item{exclude_states}{States for which we chose to
not submit a forecast for this week, as a vector with states
coded identically to how they are in the `location` columnd
of `full_diagnostics_df}

\item{output_dir}{Path to a directory in which to save
the metadata table. Only used if `prod_run = TRUE`.
Default `NULL`.}

\item{prod_run}{Is this a run for publication /
production? If so, the output will be saved in an
appropriate location as a .tsv. Boolean, default `FALSE`}
}
\value{
The metadata to be saved, as a `list`.
}
\description{
Create YAML formatted metadata file
listing states without wastewater data,
states with insufficient wastewater data,
and states for which we used the hospital
admissions only model.
}
