% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_model_file_name}
\alias{get_model_file_name}
\title{Get shortened name for writing a file with model file name}
\usage{
get_model_file_name(model_type, include_ww)
}
\arguments{
\item{model_type}{}
}
\value{
shortened model type name
}
\description{
Get shortened name for writing a file with model file name
}
