% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_model_file_path}
\alias{get_model_file_path}
\title{Get path to stan wastewater model}
\usage{
get_model_file_path(model_type)
}
\value{
the file path
}
\description{
Get path to stan wastewater model
}
