% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_params.R
\name{get_model_param_df}
\alias{get_model_param_df}
\title{Information about a ww model's parameters and generated quantities}
\usage{
get_model_param_df(x)
}
\arguments{
\item{x}{CmdStanModel or CmdStanFit object}
}
\value{
a dataframe containing the name of each parameter (and/or generated quantity)
and information about how it varies in time and across sites/labs.
}
\description{
Accepts a string describing the model name, a fitted model (output of model$sample()),
or a model object.
Returns information about the model parameters.
}
