% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_model_outputs.R
\name{get_model_run_diagnostics}
\alias{get_model_run_diagnostics}
\title{Get the parameters that were flagged for diagnostics for the model run}
\usage{
get_model_run_diagnostics(
  stan_fit_object,
  n_chains,
  rhat_threshold = 1.05,
  ess_threshold = 100
)
}
\arguments{
\item{stan_fit_object}{CmdstanR object}

\item{n_chains}{Number of chains in MCMC}

\item{rhat_threshold}{Max acceptable Rhat, above this number we flag that
parameter}

\item{ess_threshold}{this number times n chains is the minimum effective
sample size before we flag}
}
\value{
model_flags: a dataframe containing all the parameters that
are flagged for havign either high Rhat or low effective sample size
}
\description{
Get the parameters that were flagged for diagnostics for the model run
}
