% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{get_nearest_forecast_date}
\alias{get_nearest_forecast_date}
\title{Local functions that are used to pre-process our specific data for use
in the stan model
These will be used to get the data up to the point it needs to be in the
get_stan_data functions, which format the data into the stan lists that the
stan model needs.
None of these will be exported
Get forecast date from today's date}
\usage{
get_nearest_forecast_date()
}
\description{
Local functions that are used to pre-process our specific data for use
in the stan model
These will be used to get the data up to the point it needs to be in the
get_stan_data functions, which format the data into the stan lists that the
stan model needs.
None of these will be exported
Get forecast date from today's date
}
