% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_config_vals.R
\name{get_pdf_file_path}
\alias{get_pdf_file_path}
\title{get the cleaned pdf filepath}
\usage{
get_pdf_file_path(output_dir)
}
\arguments{
\item{output_dir}{}
}
\description{
get the cleaned pdf filepath
}
