% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_output_subdirectory_paths.R
\name{get_pdf_output_subdirectory}
\alias{get_pdf_output_subdirectory}
\title{Get the path to the subdirectory in which
to save summary output for a given top-level
output directory.}
\usage{
get_pdf_output_subdirectory(output_dir)
}
\arguments{
\item{output_dir}{Path to the output directory}
}
\value{
The path to the subdirectory in which
to save the summary PDFs
}
\description{
Get the path to the subdirectory in which
to save summary output for a given top-level
output directory.
}
