% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_output_metadata.R
\name{get_pipeline_metadata}
\alias{get_pipeline_metadata}
\title{Assemble pipeline metadata for a given
run, returning it as a list and
writing it to disk as a YAML file.}
\usage{
get_pipeline_metadata(
  us_run,
  dir_to_save_metadata,
  run_id,
  date_run,
  ww_data_path,
  hosp_data_dir,
  forecast_date,
  prod_run,
  model_type,
  location,
  output_dir,
  root_dir,
  ...
)
}
\arguments{
\item{us_run}{Was the model run for the US as a
whole in addition to states and territories?
If so, this will be recorded. Boolean.}

\item{dir_to_save_metadata}{Subdirectory in which to
save the resultant metadata file.}

\item{run_id}{Unique ID of the pipeline run (as a
hash generated at runtime).}

\item{date_run}{Date the pipeline was run.}

\item{hosp_data_dir}{Subdirectory in which hospital
admissions data used for the run is stored.}

\item{forecast_date}{Forecast date targeted by the
run.}

\item{prod_run}{Is this a production run? Boolean.}

\item{model_type}{What was the model type for this run?}

\item{location}{What location(s) besides the US as a whole
were run?}

\item{output_dir}{What output directory was used to
store run output.}

\item{root_dir}{root directory where pipeline is being run, used to make absolute filepaths}

\item{...}{Other keyword arguments, for compatibility
with do.call on a large parameter list.}

\item{ww_data_dir}{Subdirectory in which wastewater
data used for the run is stored.}
}
\value{
the metadata, as a list.
}
\description{
Assemble pipeline metadata for a given
run, returning it as a list and
writing it to disk as a YAML file.
}
