% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{get_plot_change_in_p_hosp_t}
\alias{get_plot_change_in_p_hosp_t}
\title{Get plot of the change in IHR over time}
\usage{
get_plot_change_in_p_hosp_t(df, figure_file_path, write_files = TRUE)
}
\arguments{
\item{df}{dataframe of filepaths}

\item{figure_file_path}{where to write the figure to}

\item{write_files}{whether or not to save the figure, default is TRUE}
}
\value{
a plot of the change in the IHR over time (IHR intercept - IHR(t))
}
\description{
Get plot of the change in IHR over time
}
