% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{get_plot_labsite_ww_draws}
\alias{get_plot_labsite_ww_draws}
\title{Plot of lab-site level wastewater draws}
\usage{
get_plot_labsite_ww_draws(
  df,
  figure_file_path,
  from_full_df = FALSE,
  write_files = TRUE
)
}
\arguments{
\item{df}{dataframe of draws or dataframe containing filepath to load in draws}

\item{figure_file_path}{path to save figure
 @param from_full_df if TRUE then df is draws object, if FALSE then df is a
dataframe of filepaths to load in the parameter object}

\item{write_files}{if TRUE then write files to specified location, if FALSE
then don't}
}
\value{
plot object of site level wastewater trajectories with observed data
overlaid
}
\description{
Plot of lab-site level wastewater draws
}
