% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{get_plot_metadata}
\alias{get_plot_metadata}
\title{Get the metadata needed for plotting}
\usage{
get_plot_metadata(df, y = NA)
}
\arguments{
\item{y}{the output your plotting, default is NA in which case colors will be
based on other variables}
}
\value{
a list with the plot color, label for y axis, and the output
description to use in the title
}
\description{
Get the metadata needed for plotting
}
