% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{get_plot_param_distribs}
\alias{get_plot_param_distribs}
\title{Get plot of parameter distributions}
\usage{
get_plot_param_distribs(
  df,
  figure_file_path,
  from_full_df = FALSE,
  write_files = TRUE,
  ...
)
}
\arguments{
\item{df}{dataframe either of parameter draws or of the filepath to obtain
the parameter draws}

\item{figure_file_path}{directory to save the figure}

\item{from_full_df}{if TRUE then df is draws object, if FALSE then df is a
dataframe of filepaths to load in the parameter object}

\item{write_files}{if TRUE then write files to specified location, if FALSE
then don't}

\item{...}{}
}
\value{
plot object of histograms of key parameters
}
\description{
Get plot of parameter distributions
}
