% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{get_plot_single_param}
\alias{get_plot_single_param}
\title{Get plot single parameter}
\usage{
get_plot_single_param(
  df,
  param_name,
  figure_file_path,
  from_full_df = FALSE,
  write_files = TRUE,
  ...
)
}
\arguments{
\item{df}{df of filepaths or of draws themselves}

\item{param_name}{the name of the parameter to plot}

\item{figure_file_path}{path to save figure}

\item{from_full_df}{whether or not df is a df of filepaths or of draws,
default = FALSE which means it is of filepaths}

\item{write_files}{whether or not to write files (default = TRUE)}

\item{...}{}
}
\description{
Get plot single parameter
}
