% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_predictive_check.R
\name{get_prior_pred_hosps_df}
\alias{get_prior_pred_hosps_df}
\title{Observables extraction from prior predictive samples}
\usage{
get_prior_pred_hosps_df(prior_pred)
}
\arguments{
\item{prior_pred}{output of prior_predictive_check, stan data object}
}
\value{
a tibble, with columns time (the time),
hosp (the hospitalization count), and
draw (the draw this came from)
}
\description{
Observables extraction from prior predictive samples
}
