% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_predictive_check.R
\name{get_prior_pred_ww_df}
\alias{get_prior_pred_ww_df}
\title{Observables extraction from prior predictive samples}
\usage{
get_prior_pred_ww_df(prior_pred, site_index)
}
\arguments{
\item{prior_pred}{output of prior_predictive_check, stan data object}

\item{site_index}{which site to extract}
}
\value{
a tibble, with columns time (the time),
log_conc (the ww log-concentration), and
draw (the draw this came from)
}
\description{
Observables extraction from prior predictive samples
}
