% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_model_outputs.R
\name{get_quantiles_on_site_level_ww}
\alias{get_quantiles_on_site_level_ww}
\title{Quantile the site level expected observed WW}
\usage{
get_quantiles_on_site_level_ww(draws)
}
\arguments{
\item{draws}{the tidy bayes output with stacked draws of all generated quantites}
}
\value{
a wide format dataframe with just the median, 50th, and 95th quantiles of
the expected WW concentation
}
\description{
Quantile the site level expected observed WW
}
