% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_random_string}
\alias{get_random_string}
\title{Get a random string of length 8 containing 3 letters, 4 digits, 1 letter}
\usage{
get_random_string(forecast_date, date_run, ww_data_path)
}
\arguments{
\item{ww_data_path}{}
}
\value{
the random string with uppercase letters
}
\description{
Get a random string of length 8 containing 3 letters, 4 digits, 1 letter
}
