% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_model_outputs.R
\name{get_raw_param_draws}
\alias{get_raw_param_draws}
\title{Get raw parameter draws}
\usage{
get_raw_param_draws(stan_output_draws, model_type, n_draws = 500, ...)
}
\arguments{
\item{...}{}
}
\value{
a long tidy dataframe with columns names:
name, t, value, draw corresponding to the value of the draw of the posterior
of that parameter
}
\description{
This concatenates all the tidy draws dataframes of each of the parameters
}
