% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{get_rt_site_level}
\alias{get_rt_site_level}
\title{Plot of site-level Rt}
\usage{
get_rt_site_level(
  df,
  figure_file_path,
  from_full_df = FALSE,
  write_files = TRUE
)
}
\arguments{
\item{df}{dataframe of either model draws or filepath to load in model draws}

\item{figure_file_path}{where to save the figure}

\item{from_full_df}{if TRUE then df is draws object, if FALSE then df is a
dataframe of filepaths to load in the parameter object}

\item{write_files}{if TRUE then write files to specified location, if FALSE
then don't}
}
\value{
plot object of R(t)s in each site with state level median overlaid
}
\description{
Plot of site-level Rt
}
