% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_model_outputs.R
\name{get_scores}
\alias{get_scores}
\title{Get scores}
\usage{
get_scores(
  quantiles,
  output_dir,
  scores_file_path = file.path("summaries", "scores"),
  grouping_vars = c("hosp_reporting_delay", "location_name", "period", "forecast_date",
    "name", "model"),
  write_files = TRUE
)
}
\arguments{
\item{write_files}{}
}
\value{
a dataframe of average scores over the grouping variables indicated
}
\description{
Reads in the quantiles and summarizes scores by the grouping variables specified
}
