% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_processing.R
\name{get_site_county_map}
\alias{get_site_county_map}
\title{Get site county mapp}
\usage{
get_site_county_map(nwss)
}
\arguments{
\item{nwss}{wastewater data}
}
\value{
a mapping from the unique county fips in the ww data to the major county
}
\description{
Get site county mapp
}
