% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_stan_data}
\alias{get_stan_data}
\title{Get stan data}
\usage{
get_stan_data(
  train_data,
  params,
  forecast_date,
  forecast_time,
  include_hosp,
  compute_likelihood,
  generation_interval,
  inf_to_hosp,
  infection_feedback_pmf,
  ...
)
}
\arguments{
\item{generation_interval}{vector of discretized probability mass indexed
starting at day 1 after infection, describing the probability of onwards
infection on each day
after infection}

\item{inf_to_hosp}{vector of discretized probability mass describing the
probability of hopsital admissions on each day of infection, indexed
starting at 0}

\item{infection_feedback_pmf}{vector of discretized probability mass
used to enforce the delay distribution on infection feedback}

\item{...}{}
}
\description{
Stan data needed for the model where we aggregate the ww data
so that we have at most one WW osbservation per day (but likely a weekly avg)
}
