% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{get_state_level_summary}
\alias{get_state_level_summary}
\title{Get summary of WW viral concentrations by state}
\usage{
get_state_level_summary(nwss_by_week)
}
\arguments{
\item{nwss_by_week}{}
}
\description{
This uses the same processing steps that biobot uses (described in:
https://github.com/biobotanalytics/covid19-wastewater-data#wastewater-data)
We make columns for the unweighted avg concentraiton across the state,
the population weighted average, the population weighted average with a
300,000 person threshold for site-specific population served, and a 3 week
rolling average of the weighted concentration with the threshold, for both
the state and national level. If there are gaps in weeks for a state,
the 3 week average is across a longer time frame. We purposely do not use
expand grid to inflate this, as we want those states with missing/ incomplete
data to be missing data when the model is fit (rather than have NA values).
}
