% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_model_outputs.R
\name{get_submission_file_path}
\alias{get_submission_file_path}
\title{Get the hardcoded submission file path}
\usage{
get_submission_file_path(output_dir)
}
\arguments{
\item{...}{}
}
\description{
Get the hardcoded submission file path
}
