% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_config_vals.R
\name{get_summarized_table}
\alias{get_summarized_table}
\title{Get summarized table with each location and}
\usage{
get_summarized_table(
  hub_submission_df,
  full_diagnostics_df,
  hosp_only_states,
  repo_file_path
)
}
\arguments{
\item{repo_file_path}{}
}
\description{
Get summarized table with each location and
}
