% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_processing.R
\name{get_testing_data}
\alias{get_testing_data}
\title{Get testing dataset}
\usage{
get_testing_data(
  data_as_of,
  forecast_date,
  location,
  calibration_time,
  forecast_time,
  hosp_reporting_delay,
  geo_type,
  hosp_data_source,
  pull_from_local,
  hosp_data_dir,
  population_data_path
)
}
\arguments{
\item{ww_data_raw}{The WW data that we need alongside
the testing hospitalization data (WW data doesn't change)}

\item{config_vars}{The config vars that tell us the
forecast date, locations, data sources, hosp reporting
delays, etc}

\item{date_as_of}{The date that we want to evaluate the
projections as of. This could be today, or could be some
weeks in the future}
}
\value{
A dataset that goes up until the forecast_date +
forecast_time and starts when the calibration started
for each location. Labels whether the data is calibration,
nowcast, or forecasted. This will be for multiple locations,
but only for a single model type, forecast date, and
hospitalizaiton reporting delay
}
\description{
Get testing dataset
}
