% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_processing.R
\name{get_training_data}
\alias{get_training_data}
\title{Get dataframe with wastewater and hospitalizaiton data for training the model}
\usage{
get_training_data(
  ww_data_raw,
  hosp_data_raw,
  forecast_date,
  calibration_time,
  location,
  hosp_reporting_delay,
  forecast_time
)
}
\arguments{
\item{forecast_time}{}
}
\value{
a dataframe of daily hospital admits, observed wastewater
concentrations, time starting at the start of the calibration period and
going til the end of the forecast period, the day of the week, period, date,
and location
}
\description{
Get dataframe with wastewater and hospitalizaiton data for training the model
}
