% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocessing.R
\name{get_weekly_summary}
\alias{get_weekly_summary}
\title{Summarize data into weekly by site}
\usage{
get_weekly_summary(nwss_subset, ww_target_type = "pcr_target_avg_conc")
}
\arguments{
\item{nwss_subset}{subsetted nwss data}
}
\value{
weekly summary of nwss data by site
}
\description{
Get one sample per site per week by averaging over the week if a site submits
more than once per week
}
