% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_processing.R
\name{get_ww_data}
\alias{get_ww_data}
\title{Get the watstewater data for a particular source
and geographic granularity}
\usage{
get_ww_data(
  ww_data_source,
  geo_type,
  ww_data_type,
  ww_target_type,
  ww_geo_type,
  ww_data_path,
  ...
)
}
\arguments{
\item{ww_data_path}{}
}
\value{
dataframe with observed viral concentrations at the geographic
level specified and weekly temporal granularity. Dates correspond to the
middle of the week. Will be used to map to hospitalization data.
}
\description{
Wrapper function around the worker functions.
Outputs a different dataframe depending on if
the pipeline is site level (ww_geo_type != geo_type)
or state level
}
