% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecasttools.R
\name{loc_abbr_to_flusight_code}
\alias{loc_abbr_to_flusight_code}
\title{Convert a two-letter location abbreviation to a
two-character FluSight location code}
\usage{
loc_abbr_to_flusight_code(abbrs)
}
\arguments{
\item{abbrs}{vector of USPS two letter name abbreviations}
}
\value{
vector of the same length recoded as flusight
location codes
}
\description{
Given a vector of state/territory two-letter
USPS short names (e.g. MA, TX, PR), return
the corresponding FluSight challenge location
code (legacy FIPS code for states and territories,
`US` for the US).
}
\seealso{
[loc_flusight_code_to_abbr()]
}
