% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EDA_utils.R
\name{log_standardize}
\alias{log_standardize}
\title{As standardize, but on log scale, possibly removing zeros (or near-zeros)}
\usage{
log_standardize(x, use_mad = FALSE, zero_threshold = 1e-08)
}
\description{
As standardize, but on log scale, possibly removing zeros (or near-zeros)
}
