% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_pps_data.R
\name{make_pps_data}
\alias{make_pps_data}
\title{Gets generated quantites from a fitted ww model and formats for stan}
\usage{
make_pps_data(mcmc_fit, real_data, config, draw = NULL)
}
\arguments{
\item{mcmc_fit}{CmdStanMCMC object}

\item{real_data}{the stan data object which produced the mcmc_fit}

\item{config}{the output of get_config_vals() which produced the mcmc_fit}

\item{draw}{vector, which draws from the MCMC object do we want in the result?
If NULL, uses all. This can be a very large data frame!}
}
\value{
a list, each of which is a list of formatted stan data
}
\description{
This function takes in a fitted wastewater model and some of the inputs which produced
it and returns data which can be used to fit the model to the posterior
predictive distribution of datasets.
}
