% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delay_distribs.R
\name{make_reporting_delay_pmf}
\alias{make_reporting_delay_pmf}
\title{Make reporting delay pmf}
\usage{
make_reporting_delay_pmf(incubation_period_pmf, hospital_onset_delay_pmf)
}
\arguments{
\item{hospital_onset_delay_pmf}{}
}
\value{
convolution of incubation period and sympton onset to hospital
admission pmf
}
\description{
Convolve the incubation period pmf with the symptom to hospital admission pmf
and normalize
}
