% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{param_df}
\alias{param_df}
\title{Input parameters used to generate `example_df`}
\format{
## param_df
A data frame with 305 rows and 4 columns
\describe{
  \item{name}{Parameter name}
  \item{true_value}{ The value that was assigned to this parameter in the
  data generation process}
  \item{index_rows}{If the data is dynamic, the other indexing }
  \item{index_cols}{If the data is dynamic the indexing (e.g. time in
  days or weeks) }
}
}
\source{
generate_simulated_data.R
}
\usage{
param_df
}
\description{
A few key parameters that were created by the default arguments in
`generate_simulated_data.R`
}
\keyword{datasets}
