% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_model_outputs.R
\name{pivot_data_for_plotting}
\alias{pivot_data_for_plotting}
\title{Pivot the quantiles output for easy plotting}
\usage{
pivot_data_for_plotting(quantiles)
}
\arguments{
\item{quantiles}{a long dataframe of quantiles}
}
\value{
a wide dataframe of quantiles with just the quantiles we want for plotting
}
\description{
Pivot the quantiles output for easy plotting
}
