% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_combined_data}
\alias{plot_combined_data}
\title{Plot the WW and hospitalization data together}
\usage{
plot_combined_data(
  comb,
  figure_file_path,
  days_to_show_calib_data = 90,
  write_files = TRUE,
  ...
)
}
\arguments{
\item{comb}{dataframe for a single location to plot}

\item{figure_file_path}{higher level directory where data from this pipeline
will be saved}

\item{days_to_show_calib_data}{number of days before last hospital admissions
day to put on plot, default is 90}

\item{write_files}{TRUE if write to file location}
}
\value{
a plot for a single location the scaled WW versus hospital admissions
}
\description{
Plot the WW and hospitalization data together
}
