% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prior_predictive_check.R
\name{prior_predictive_check}
\alias{prior_predictive_check}
\title{Draws data under the model's prior.}
\usage{
prior_predictive_check(model, real_data, config, ndraws, ...)
}
\arguments{
\item{model}{a CmdStanModel object defining the model}

\item{real_data}{observed data, formated for stan, needed to match, for example, the number
of wastewater sites}

\item{config}{the output of get_config_vals()}

\item{ndraws}{integer, number of draws from the prior predictive distribution to return}

\item{...}{further arguments passed to model$sample(), e.g. # chains}

\item{mclapply_cores}{integer specifying number of threads to use for
posterior-predictive-analysis-level parallelization, separate from stan parallelization}
}
\value{
a list, each of which is a list of formatted stan data
}
\description{
This function takes in a model object, observed data, and the configuration file used to
produce the real-data stan data object. It then runs the model under the prior and
returns the observable (generated quantities) for some number of draws.
}
