% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_model_outputs.R
\name{read_diagnostics_df}
\alias{read_diagnostics_df}
\title{Load in and combine all the diagnostic tables}
\usage{
read_diagnostics_df(df_of_filepaths)
}
\arguments{
\item{df_of_filepaths}{dataframe containing filepath indicating where
model outputs are saved}
}
\value{
a combined list of all the diagnostics across locations and model
runs
}
\description{
Load in and combine all the diagnostic tables
}
