% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_diagnostic_report.R
\name{render_diagnostic_report}
\alias{render_diagnostic_report}
\title{Render diagnostic report}
\usage{
render_diagnostic_report(
  output,
  params,
  report_rmd = system.file("diagnostic_report_non_targets.Rmd", package =
    "cfaforecastrenewalww", mustWork = TRUE),
  ...
)
}
\arguments{
\item{output}{path to output .html}

\item{params}{list of parameters for the Rmd}

\item{report_rmd}{path to Rmarkdown skeleton}

\item{...}{additional parameters passed to \code{rmarkdown::render}}
}
\description{
Render diagnostic report
}
