% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_model_outputs.R
\name{save_to_pdf}
\alias{save_to_pdf}
\title{Save a single pdf of plots}
\usage{
save_to_pdf(
  list_of_plots,
  type_of_output,
  pdf_file_path,
  forecast_date,
  model_type,
  n_row = 3,
  n_col = 2,
  ...
)
}
\arguments{
\item{list_of_plots}{the output from targets of a list of ggplot objects}

\item{type_of_output}{what the pdf is a plot of}

\item{pdf_file_path}{where we want to save the pdf within the output file path}

\item{forecast_date}{date of forecast}

\item{model_type}{which model we are running}

\item{n_row}{number of rows of figures on a page}

\item{n_col}{number of columns of figures on a page}

\item{...}{}

\item{output_dir}{where to save the output}
}
\description{
Saves pdf containing the concatenated figures
}
