% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delay_distribs.R
\name{simulate_double_censored_pmf}
\alias{simulate_double_censored_pmf}
\title{Simulate daily double censored PMF. From {epinowcast}:
https://package.epinowcast.org/dev/reference/simulate_double_censored_pmf.html}
\usage{
simulate_double_censored_pmf(
  max,
  fun_primary = stats::runif,
  primary_args = list(),
  fun_dist = stats::rlnorm,
  dist_args = list(...),
  n = 1e+06,
  ...
)
}
\arguments{
\item{max}{Maximum value for the computed CDF. If not specified, the maximum
value is the maximum simulated delay.}

\item{fun_primary}{Primary distribution function (default is \code{runif}).}

\item{primary_args}{List of additional arguments to be passed to the primary
distribution function.}

\item{fun_dist}{Distribution function to be added to the primary (default is
\code{rlnorm}).}

\item{dist_args}{List of additional arguments to be passed to the
distribution function.}

\item{n}{Number of simulations (default is 1e6).}

\item{...}{Additional arguments to be passed to the distribution function.
This is an alternative to `dist_args`.}
}
\value{
A numeric vector representing the PMF.
}
\description{
This function simulates the probability mass function of a  daily
double-censored process. The process involves two distributions: a primary
distribution which represents the censoring process for the primary event
and another distribution (which is offset by the primary).
}
\details{
Based off of:
https://www.medrxiv.org/content/10.1101/2024.01.12.24301247v1
}
\examples{
simulate_double_censored_pmf(10, meanlog = 0, sdlog = 1)
}
\concept{modelmodulehelpers}
