% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{site_level_obs_inits}
\alias{site_level_obs_inits}
\title{Get initialization function for stan for site-level observation error model}
\usage{
site_level_obs_inits(train_data, params, stan_data)
}
\value{
the initialization function
}
\description{
Get initialization function for stan for site-level observation error model
}
\keyword{internal}
