% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecasttools.R
\name{soql_is_in}
\alias{soql_is_in}
\title{Return a [soql::soql_where()] construct
for a given column being in a list of values}
\usage{
soql_is_in(soql_list, column, match_values)
}
\arguments{
\item{soql_list}{A `soql` query object, which
can be piped in. If one hasn't been
created yet, use or pipe in [soql::soql()].}

\item{column}{The column to filter on}

\item{match_values}{A vector of values that column
must match}
}
\value{
A new soql object with the filter added,
for use in other functions.
}
\description{
Return a [soql::soql_where()] construct
for a given column being in a list of values
}
