% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EDA_utils.R
\name{standardize}
\alias{standardize}
\title{x -> (x - location(x))/dispersion(x)
location is mean, dispersion is SD or MAD}
\usage{
standardize(x, use_mad = FALSE)
}
\description{
x -> (x - location(x))/dispersion(x)
location is mean, dispersion is SD or MAD
}
