% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{state_agg_inits}
\alias{state_agg_inits}
\title{Get initialization function for stan for aggregated model}
\usage{
state_agg_inits(train_data, params, stan_data)
}
\value{
the initialization function
}
\description{
Get initialization function for stan for aggregated model
}
\keyword{internal}
