% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_model_outputs.R
\name{summarize_scores}
\alias{summarize_scores}
\title{Summarize scores}
\usage{
summarize_scores(scores)
}
\arguments{
\item{scores}{}
}
\value{
wide dataframe of average scores across time
}
\description{
Takes the scores grouped by location, period, model, data type (name),
and hosp_reporting delay and averages over them. Used to get summaries to
compare across models and scenarios over time
}
