% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{time_varying_conc_inits}
\alias{time_varying_conc_inits}
\title{Get initialization function for stan for time-varying concentration model}
\usage{
time_varying_conc_inits(train_data, params, stan_data)
}
\value{
the initialization function
}
\description{
Get initialization function for stan for time-varying concentration model
}
\keyword{internal}
