% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecasttools.R
\name{trajectories_to_quantiles}
\alias{trajectories_to_quantiles}
\title{Aggregate individual trajectory
timeseries or forecasts to quantile
timeseries or forecasts}
\usage{
trajectories_to_quantiles(
  trajectories,
  quantiles = c(0.01, 0.025, seq(0.05, 0.95, 0.05), 0.975, 0.99),
  timepoint_cols = "timepoint",
  value_col = "value",
  quantile_value_name = "quantile_value",
  quantile_level_name = "quantile_level",
  id_cols = NULL
)
}
\arguments{
\item{trajectories}{tidy data frame or tibble
of trajectories}

\item{quantiles}{quantiles to output for each
timepoint (default the FluSight 2023 quantiles:
`c(0.01, 0.025, seq(0.05, 0.95, 0.05), 0.975, 0.99)`}

\item{timepoint_cols}{name of the column(s) in`trajectories`
that identifies unique timepoints. Default `timepoint`.}

\item{value_col}{name of the column in `trajectories`
with the trajectory values (for which we wish to
compute quantiles), e.g. `hosp`, `weekly_hosp`, `cases`,
etc. Default `value`.}

\item{quantile_value_name}{What to name
the column containing quantile values in
the output table. Default `"quantile_value"`}

\item{quantile_level_name}{What to name
the column containing quantile levels in
the output table. Default `"quantile_level"`}

\item{id_cols}{additional id columns in
`trajectories` to group by before aggregating,
e.g. a `location` column if `trajectories` contains
trajectories over the same time period for multiple
locations, such as different US States and Territories.
If NULL, ignored. Default NULL.}
}
\description{
Given a tidy data frame of
trajectories, aggregate it to
a quantile timeseries for the
given quantile values
}
