% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_params.R
\name{validate_paramlist}
\alias{validate_paramlist}
\title{Validate a parameter list}
\usage{
validate_paramlist(paramlist)
}
\arguments{
\item{paramlist}{The parameter list to validate}
}
\value{
the parameter list, on success,
or raise an error
}
\description{
Validate a parameter list
}
