% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{ww_model}
\alias{ww_model}
\title{This code was adapted from code written
(under an MIT license) as part of the `epinowcast`
package (https://github.com/epinowcast/epinowcast)
ww_model}
\usage{
ww_model(
  model,
  include = system.file("stan", package = "cfaforecastrenewalww"),
  compile = TRUE,
  threads = FALSE,
  target_dir = tempdir(),
  stanc_options = list(),
  cpp_options = list(),
  verbose = TRUE,
  ...
)
}
\arguments{
\item{...}{}
}
\description{
This function compiles the stan model, and is written to include the 'stan'
folder. Within each stan file, to include the functions, use #include
functions/{your_function_file}.stan
}
