% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EDA_utils.R
\name{zoom_boxplot_y}
\alias{zoom_boxplot_y}
\title{Get y range required to zoom a boxplot in to ignore outliers
df is the data, x the grouping variable, y the variable to be boxplotted}
\usage{
zoom_boxplot_y(df, x, y)
}
\description{
Get y range required to zoom a boxplot in to ignore outliers
df is the data, x the grouping variable, y the variable to be boxplotted
}
