% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfaforecastrenewalww-package.R
\docType{package}
\name{cfaforecastrenewalww-package}
\alias{cfaforecastrenewalww}
\alias{cfaforecastrenewalww-package}
\title{cfaforecastrenewalww: Wastewater informed COVID-19 hospital admissions forecasting}
\description{
Semi-mechanistic renewal approach jointly calibrating to state-level hospital admissions and site-level wastewater concentrations.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/cdcgov/wastewater-informed-covid-forecasting/}
  \item Report bugs at \url{https://github.com/cdcgov/wastewater-informed-covid-forecasting/issues/}
}

}
\author{
\strong{Maintainer}: Kaitlyn Johnson \email{uox1@cdc.gov} (\href{https://orcid.org/0000-0001-8011-0012}{ORCID})

Authors:
\itemize{
  \item Sam Abbott \email{contact@samabbott.co.uk} (\href{https://orcid.org/0000-0001-8057-8037}{ORCID})
  \item Zachary Susswein \email{utb2@cdc.gov}
  \item Andrew Magee \email{rzg0@cdc.gov}
  \item Dylan Morris \email{dylan@dylanhmorris.com} (\href{https://orcid.org/0000-0002-3655-406X}{ORCID})
  \item Scott Olesen \email{ulp7@cdc.gov}
  \item Damon Bayer \email{xum8@cdc.gov}
}

Other contributors:
\itemize{
  \item George Vega Yon \email{g.vegayon@gmail.com} (\href{https://orcid.org/0000-0002-3171-0844}{ORCID}) [contributor]
}

}
\keyword{internal}
