% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compile_model.R
\name{compile_model}
\alias{compile_model}
\title{Compile a stan model while pointing at the package default
include directory (\code{stan}) for #include statements}
\usage{
compile_model(
  model_filepath,
  include_paths = system.file("stan", package = "cfaforecastrenewalww"),
  threads = FALSE,
  target_dir = tempdir(),
  stanc_options = list(),
  cpp_options = list(),
  verbose = TRUE,
  ...
)
}
\arguments{
\item{model_filepath}{path to .stan file defining the model}

\item{include_paths}{path(s) to directories to search for files
specified in #include statements. Passed to \code{\link[cmdstanr:cmdstan_model]{cmdstanr::cmdstan_model()}}.
Defaults to the \code{stan} subdirectory of the installed
\code{cfaforecastrenewalww} package.}

\item{threads}{Number of threads to use in model compilation,
as an integer. Passed to \code{\link[cmdstanr:cmdstan_model]{cmdstanr::cmdstan_model()}}.
Default \code{FALSE} (use single-threaded compilation).}

\item{target_dir}{Directory in which to save the compiled
stan model binary. Passed as the \code{dir} keyword argument to
\code{\link[cmdstanr:cmdstan_model]{cmdstanr::cmdstan_model()}}. Defaults to a temporary directory
for the R session (the output of \code{\link[=tempdir]{tempdir()}}).}

\item{stanc_options}{Options for the stan compiler passed to
\code{\link[cmdstanr:cmdstan_model]{cmdstanr::cmdstan_model()}}, as a list. See that function's
documentation for more details. Default \code{list()} (use default
options).}

\item{cpp_options}{Options for the C++ compiler passed to
\code{\link[cmdstanr:cmdstan_model]{cmdstanr::cmdstan_model()}}, as a list. See that function's
documentation for more details. Default \code{list()} (use default
options).}

\item{verbose}{Write detailed output to the terminal while
executing the function? Boolean, default \code{TRUE}.}

\item{...}{Additional keyword arguments passed to
\code{\link[cmdstanr:cmdstan_model]{cmdstanr::cmdstan_model()}}.}
}
\value{
The resulting \code{cmdstanr} model object, as the output
of \code{\link[cmdstanr:cmdstan_model]{cmdstanr::cmdstan_model()}}.
}
\description{
This function reads in and optionally compiles a Stan model.
It is written to search the installed package \code{stan} directory
for additional source files to include. Within each stan file,
use #include {path to your file with the \code{stan} directory}.stan
}
\details{
The code for this function has been adapted
from code written (under an MIT license) as part of
the \href{https://github.com/epinowcast/epinowcast}{\code{epinowcast}}
R package.
}
