% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_df}
\alias{example_df}
\title{Input dataframe for fitting the site-level infection dynamics wastewater
model}
\format{
## example_df
A data frame with 635 rows and 13 columns
\describe{
  \item{t}{The time index in days}
  \item{lab_wwtp_unique_id}{The unique identifier of the wasteawter site-lab
   combination (1-5) in this instance,
  but would vary for the number of unique combinations in your inptu dataset
  if trying to construct something mirroring
  this structure}
  \item{log_conc}{The log scaled viral genomes per mL collected from the
  lab-site specified on day date.
  NAs indicate days where a sample wasn't reported in that lab-site.}
  \item{date}{Sample collection date, formatted as YYYY-MM-DD}
  \item{lod_sewage}{The log scaled limit of detection reported by the
  lab-site on that date}
  \item{below_LOD}{An indicator (0,1) indicating whether the wastewater
  observation on that date is
   above (0) or below (1) the LOD. NAs for all days with wastewater
   observations in that lab-site}
  \item{daily_hosp_admits}{The number of individuals admitted to the
  hospital on that date,
  available as of the forecast date. Note this is repeated for each site-lab.}
  \item{daily_hosp_admits_for_eval}{The number of individuals admitted to
  the hospital on that date,
  available retrospectively (so including after the forecast date).
  Note this is repeated for each site-lab}
  \item{pop}{The number of people in the "state" or whatever population
  contributes to the daily hospital admissions}
  \item{forecast_date}{The date the forecast is made, formatted as YYYY-MM-DD}
  \item{hosp_calibration_time}{The duration of time in days to calibrate the
   model to hospital admissions data}
  \item{site}{The unique identifier for the site (also referred to as a
  wastewater catchment area). There can be more than
  one lab per site if samples are sent to different/overlapping labs.}
  \item{ww_pop}{The population size of the site (wastewater catchment area)}
}
}
\source{
generate_simulated_data.R
}
\usage{
example_df
}
\description{
A dataset containing the daily hospital admissons, observed wastewater
concentrations and other attributes needed for the model to be fit.
Generated via the defaults in `generate_simulated_data.R`
 The variables are as follows:
}
\keyword{datasets}
