% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{get_df_of_filepaths}
\alias{get_df_of_filepaths}
\title{Get dataframe of filepaths}
\usage{
get_df_of_filepaths(
  location,
  model_type,
  forecast_date,
  date_run,
  run_id,
  output_file_path
)
}
\arguments{
\item{output_file_path}{}
}
\description{
Get dataframe of filepaths
}
