% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_model.R
\name{get_df_of_filepaths}
\alias{get_df_of_filepaths}
\title{This code was adapted from code written
(under an MIT license) as part of the `epinowcast`
package (https://github.com/epinowcast/epinowcast)
Get dataframe of filepaths}
\usage{
get_df_of_filepaths(output_dir, forecast_date, location, model_type)
}
\arguments{
\item{model_type}{}
}
\description{
This code was adapted from code written
(under an MIT license) as part of the `epinowcast`
package (https://github.com/epinowcast/epinowcast)
Get dataframe of filepaths
}
