% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_model_outputs.R
\name{get_diagnostics}
\alias{get_diagnostics}
\title{Get a data frame of data and model fit diagnostics}
\usage{
get_diagnostics(
  stan_fit_object,
  train_data,
  location,
  model_type,
  forecast_date,
  ebmfi_tolerance = 0.2,
  divergences_tolerance = 0.01,
  p_high_rhat_tolerance = 0.05,
  max_tree_depth_tol = 0.01,
  ...
)
}
\arguments{
\item{stan_fit_object}{The cmdstan object}

\item{train_data}{The dataframe containing hospital admissions and
wastewater data for callibration of the model}

\item{location}{The location the model is being run on}

\item{model_type}{The type of model}

\item{forecast_date}{The date of the forecast}

\item{ebmfi_tolerance}{tolerance for mean EBMFI, default of 0.2}

\item{divergences_tolerance}{tolerance for percent of iterations that
are divergent from an individual chain}

\item{p_high_rhat_tolerance}{tolerance for probability of high p_hat,
default 0.05}

\item{max_tree_depth_tol}{tolerance for number of draws that
hit the maximum tree depth, default to 0.01}

\item{...}{}
}
\value{
Returns a dataframe containing diagnostic summaries
}
\description{
Get a data frame of data and model fit diagnostics
}
