% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_config_vals.R
\name{get_figure_file_path}
\alias{get_figure_file_path}
\title{Get the figure file path}
\usage{
get_figure_file_path(output_dir)
}
\arguments{
\item{output_dir}{}
}
\description{
Get the figure file path
}
