% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_config_vals.R
\name{get_metadata_yaml}
\alias{get_metadata_yaml}
\title{Get metadata to list all states with no wastewater or insufficient wastewater
data}
\usage{
get_metadata_yaml(full_diagnostics_df, repo_file_path, hosp_only_states, ...)
}
\arguments{
\item{hosp_only_states}{}
}
\description{
Get metadata to list all states with no wastewater or insufficient wastewater
data
}
