% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_params.R
\name{get_params}
\alias{get_params}
\title{Get parameters for model run}
\usage{
get_params(param_file)
}
\arguments{
\item{param_file}{Path to a \code{.toml} file defining
parameter values.}
}
\value{
a dataframe with numeric values for parameter values passed to the
model
}
\description{
Get parameters for model run
}
