% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_config_vals.R, R/process_model_outputs.R
\name{get_pdf_file_path}
\alias{get_pdf_file_path}
\title{get the cleaned pdf filepath}
\usage{
get_pdf_file_path(output_file_path, forecast_date, date_run)

get_pdf_file_path(output_file_path, forecast_date, date_run)
}
\arguments{
\item{date_run}{}
}
\description{
get the cleaned pdf filepath

Get the hardcoded file path for pdfs per the read me
}
