% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_config_vals.R
\name{get_pipeline_metadata}
\alias{get_pipeline_metadata}
\title{Write pipeline metadata to a YAML}
\usage{
get_pipeline_metadata(
  us_run,
  path_to_save_metadata,
  run_id,
  date_run,
  ww_data_path,
  hosp_data_dir,
  forecast_date,
  prod_run,
  model_type,
  location,
  output_dir,
  root_dir,
  ...
)
}
\arguments{
\item{us_run}{US run?}

\item{root_dir}{root directory where pipeline is being run, used to make absolute filepaths}
}
\value{
the metadata
}
\description{
Write pipeline metadata to a YAML
}
