% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_config_vals.R, R/process_model_outputs.R
\name{get_submission_file_path}
\alias{get_submission_file_path}
\title{Get the submission file path fior hub csv}
\usage{
get_submission_file_path(output_dir, forecast_date, date_run)

get_submission_file_path(output_dir, forecast_date, date_run)
}
\arguments{
\item{date_run}{}

\item{...}{}
}
\description{
Get the submission file path fior hub csv

Get the hardcoded submission file path
}
