% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_model_outputs.R
\name{get_submission_filepath_df}
\alias{get_submission_filepath_df}
\title{Get a consolidated dataframe with each location and its corresponding
file path for the necesary outputs.}
\usage{
get_submission_filepath_df(
  prod_model_type,
  loc_model_map,
  df_of_filepaths_inf_dyn = NULL,
  df_of_filepaths_hosp_only = NULL,
  df_of_filepaths_agg = NULL
)
}
\arguments{
\item{prod_model_type}{the model type we want to submit, default is
"site-level infection dynamics"}

\item{loc_model_map}{two column dataframe with location and desired model type}

\item{df_of_filepaths_inf_dyn}{the output filepaths resulting after running
the infection dynamics model targets, default is NULL}

\item{df_of_filepaths_hosp_only}{the output filepaths resulting after running
the hospital admissions only model targets, default is NULL}

\item{df_of_filepaths_agg}{the output filepaths resulting after running the
national aggreated model target, default is NULL}
}
\value{
a dataframe of length of the number of locations (will message if
not = 53) containing the model type and filepaths for the model outputs needed
for the hub submission
}
\description{
Get a consolidated dataframe with each location and its corresponding
file path for the necesary outputs.
}
