% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_model_outputs.R
\name{get_summary_stats}
\alias{get_summary_stats}
\title{Get sumamry stats of diagnostics}
\usage{
get_summary_stats(df)
}
\arguments{
\item{df}{}
}
\description{
Takes the long dataframes of diagnostics by location and outputs summaries
}
