% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{param_df}
\alias{param_df}
\title{Example input parameters}
\format{
## param_df
A data frame with 305 rows and 4 columns
\describe{
  \item{name}{Parameter name}
  \item{true_value}{}
  \item{index_rows}{}
  \item{index_cols}{}
}
}
\source{
wwww.wherever.com
}
\usage{
param_df
}
\description{
SOME HIGH-LEVEL DESCRIPTION OF THE DATA, INCLUDING HOW DERIVED
}
\keyword{datasets}
