% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_predictive_analysis.R
\name{posterior_predictive_analysis}
\alias{posterior_predictive_analysis}
\title{Runs posterior-predictive analyses and returns summaries of performance}
\usage{
posterior_predictive_analysis(
  model,
  real_data,
  config,
  draw = NULL,
  real_data_fit = NA,
  ci_width = 0.89,
  mclapply_cores = NA,
  ...
)
}
\arguments{
\item{model}{a CmdStanModel object defining the model}

\item{real_data}{observed data, formated for stan}

\item{config}{the output of get_config_vals()}

\item{draw}{vector, which draws from the MCMC object do we want to run PPS analyses for?
If NULL, runs one analysis per sample. This can take a long time and a lot of space!}

\item{real_data_fit}{CmdStanMCMC object, optional argument providing a
previous real-data model fit}

\item{mclapply_cores}{integer specifying number of threads to use for
posterior-predictive-analysis-level parallelization, separate from stan parallelization}

\item{...}{further arguments passed to model$sample(), e.g. # chains}

\item{pps_mcmc_fits}{list of CmdStanMCMC objects, optional argument providing
previously-run pps analyses (for re-summarizing)}
}
\value{
a list, each of which is a list of formatted stan data
}
\description{
This function takes in a model object, the observed data, and the configuration file used to
produce the real-data stan data object. It then fits the model, extracts generated quantities
from some number of those fits, and runs new analyses on the simulated data. It then returns
summaries of estimation performance.
}
